Bloom-backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
python-dotenv==1.0.0
google-generativeai
google-cloud-aiplatform
requests
google-adk==1.3.0
